Fix OS X support and clean up getMendeleyFolder.
The Mendeley Folder on OS X is located in ~/Library/Application
Support/Mendeley.

Store the complete path in hdir and return it once at the end.

os.path.join is only useful if you are unaware of the platform and
hence need to join the directories with os.sep. If you already know
the platform there's no need and can use os.expanduser in one go.